hboot: new command
This command will prompt for a username and password, attempt to unlock
the user's keyring with it, then store their password in the Linux
keyctl for later use.

This is intended for use in an initramfs to provide early login support.
Future patches will update himitsud to fetch the password from keyctl
to start in an unlocked state. This will allow the user to enter their
username and password once at login, optionally spin up himitsud in the
initramfs to fetch FDE passwords, then auto-login once booted and start
himitsud without any further passwords required.