query: do not allow duplicate keys
There are less edge cases in querying entries when disallowing duplicate
keys in a query and subsequently an entry.

For example to query an entry with the subset `a=a a=b`, the syntax
should ideally support different matching rules for entries that have
either value in `a` or have multiple `a` keys with specified values and
other cases like having exact amounts of duplicates with specific
values.

Disallowing duplicates will avoid such cases.