add remember options as module
My inital approach was an enum with session, skip and refuse + a
timestamp type. It was tedious to handle since you had to have a match +
switch. I didn't find proper namings of those types to include them in
himitsu::client that's why I've decided to create an extra module.