secstore: don't bail when deleting missing file
Due to some unspecified error, my database reached a state where an
entry's file was missing, but it was still present in the index.

At this point, attempting to delete it bails when attempting to delete
the missing file. Dumping the entire database doesn't work either, since
that tries to read the missing file.

When deleting an entry, if the file does not exist, simply print a
warning and continue removing it from the index.